Can view blogs while not logged in

+++ controllers/blogController.js
@@ -16,7 +16,7 @@ const blog_details = (req, res) => {
   const id = req.params.id;
   Blog.findById(id)
     .then(result => {
-      res.render('details', { blog: result, title: 'Blog Details', user: req?.user?.username, user_id: req.user._id });
+      res.render('details', { blog: result, title: 'Blog Details', user: req?.user?.username, user_id: req?.user?._id });
     })
     .catch(err => {
       console.log(err);
@@ -25,10 +25,12 @@ const blog_details = (req, res) => {
 }
 
 const blog_create_get = (req, res) => {
+  if (!req.isAuthenticated()) res.redirect('/login');
   res.render('create', { title: 'Create a new blog', user: req?.user?.username });
 }
 
 const blog_create_post = (req, res) => {
+  if (!req.isAuthenticated()) res.redirect('/login');
   const blog = new Blog({title: req.body.title, snippet: req.body.snippet, body: req.body.body, createdBy: req.user.username, createdById: req.user._id});
   const tags = req.body.tags_combined.split(',');
   for (i = 0; i < tags.length; i++) {
@@ -65,6 +67,7 @@ const blog_tag_search = (req, res) => {
 }
 
 const blog_edit = (req, res) => {
+  if (!req.isAuthenticated()) res.redirect('/login');
   const id = req.params.id;
   Blog.findById(id)
     .then(result => {
@@ -80,6 +83,7 @@ const blog_edit = (req, res) => {
 }
 
 const blog_edit_post = (req, res) => {
+  if (!req.isAuthenticated()) res.redirect('/login');
   const id = req.params.id;
 
   Blog.findById(id)
@@ -114,13 +118,24 @@ const blog_edit_post = (req, res) => {
 }
 
 const blog_delete = (req, res) => {
+  if (!req.isAuthenticated()) res.redirect('/login');
   const id = req.params.id;
-  Blog.findByIdAndDelete(id)
+  Blog.findById(id)
     .then(result => {
-      res.json({ redirect: '/blogs' });
+      if (req.user._id == result.createdById) {
+        Blog.findByIdAndDelete(id)
+        .then(result => {
+          res.json({ redirect: '/blogs' });
+        })
+        .catch(err => {
+          console.log(err);
+        });
+      } else
+        res.redirect('/blogs/'+id);
     })
     .catch(err => {
       console.log(err);
+      res.render('404', { title: 'Blog not found', user: req?.user?.username });
     });
 }
